Move redis client into stream handler
